// ordertracker-angular/src/app/api/ordertracker-backend/model/product.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {Image} from './Image';

export interface Product {
  id?: number;
  name?: string;
  price?: number;
  isVisible?: boolean;
  images?: Array<Image>;
}